This file is an example of what we call a 'bad' matlab code, we go on to show how just by good MATLAB practices one could drastically obtain speedup.
This code is simple in structure, its a MATLAB script, which obtains parameters of laser cavity from an external file called laser_param.m, uses them to calculate steady state values for given injection ratio and phase( this is the handle we have on detuning frequency as far as simulations concerned). These are then fed into the native ODE45 solver of the MATLAB. With steady state values as initial condition. The input data is just a square wave which has been filtered using a Bessel filter to smooth it out so that ODE45 doesn't have to deal with stiff regions. This is used to create a anonymous function 'f' which is nothing but a lookup table using interp1 and then this look table is used when calling leqn_eqn_solver_fun3. Which has the shared global variables( which are steady state and user defined parameters in the previous script). This contains the function code for the laser rate equation in the form of a coupled differential equation.